Refactor Moved inner class
